[Fix] Hopefully make compilers complaining about unsatisfiable condition happy (broke in previous commit)
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5174 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_it.cpp
@@ -2014,7 +2014,7 @@ void CSoundFile::SaveExtendedSongProperties(FILE* f) const
 #define WRITEMODULARHEADER(code, fsize) \
 	{ \
 		mpt::IO::WriteIntLE<uint32>(f, code); \
-		MPT_ASSERT(fsize <= uint16_max); \
+		MPT_ASSERT(static_cast<size_t>(fsize) <= uint16_max); \
 		const uint16 _size = fsize; \
 		mpt::IO::WriteIntLE<uint16>(f, _size); \
 	}
